+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/AppContext.kt
@@ -15,12 +15,9 @@ import com.facebook.react.uimanager.common.UIManagerType
 import expo.modules.adapters.react.NativeModulesProxy
 import expo.modules.core.errors.ContextDestroyedException
 import expo.modules.core.interfaces.ActivityProvider
-import expo.modules.interfaces.camera.CameraViewInterface
 import expo.modules.interfaces.constants.ConstantsInterface
-import expo.modules.interfaces.font.FontManagerInterface
 import expo.modules.interfaces.imageloader.ImageLoaderInterface
 import expo.modules.interfaces.permissions.Permissions
-import expo.modules.interfaces.taskManager.TaskManagerInterface
 import expo.modules.kotlin.activityresult.ActivityResultsManager
 import expo.modules.kotlin.activityresult.DefaultAppContextActivityResultCaller
 import expo.modules.kotlin.defaultmodules.ErrorManagerModule
@@ -37,7 +34,10 @@ import expo.modules.kotlin.providers.CurrentActivityProvider
 import expo.modules.kotlin.runtime.MainRuntime
 import expo.modules.kotlin.runtime.Runtime
 import expo.modules.kotlin.runtime.WorkletRuntime
-import expo.modules.kotlin.services.ServicesProvider
+import expo.modules.kotlin.services.AppDirectoriesService
+import expo.modules.kotlin.services.FilePermissionService
+import expo.modules.kotlin.services.Service
+import expo.modules.kotlin.services.ServicesRegistry
 import expo.modules.kotlin.tracing.trace
 import kotlinx.coroutines.CoroutineName
 import kotlinx.coroutines.CoroutineScope
@@ -51,8 +51,7 @@ import java.lang.ref.WeakReference
 class AppContext(
   modulesProvider: ModulesProvider,
   val legacyModuleRegistry: expo.modules.core.ModuleRegistry,
-  reactContextHolder: WeakReference<ReactApplicationContext>,
-  private val servicesProvider: ServicesProvider
+  reactContextHolder: WeakReference<ReactApplicationContext>
 ) : CurrentActivityProvider {
   // The main context used in the app.
   // Modules attached to this context will be available on the main js context.
@@ -101,10 +100,14 @@ class AppContext(
 
   val registry = ModuleRegistry(this.weak())
 
+  val services = ServicesRegistry(this.weak())
+
   internal var legacyModulesProxyHolder: WeakReference<NativeModulesProxy>? = null
 
   private val activityResultsManager = ActivityResultsManager(this)
-  internal val appContextActivityResultCaller = DefaultAppContextActivityResultCaller(activityResultsManager)
+  internal val appContextActivityResultCaller = DefaultAppContextActivityResultCaller(
+    activityResultsManager
+  )
 
   init {
     requireNotNull(reactContextHolder.get()) {
@@ -112,6 +115,11 @@ class AppContext(
     }.apply {
       addLifecycleEventListener(reactLifecycleDelegate)
       addActivityEventListener(reactLifecycleDelegate)
+
+      services.register<FilePermissionService>()
+      services.register<AppDirectoriesService>()
+
+      services.register(modulesProvider.getServices())
 
       // Registering modules has to happen at the very end of `AppContext` creation. Some modules need to access
       // `AppContext` during their initialisation, so we need to ensure all `AppContext`'s
@@ -150,6 +158,11 @@ class AppContext(
   }
 
   /**
+   * Returns a service implementing given interface.
+   */
+  inline fun <reified T : Service> service(): T? = services.service<T>()
+
+  /**
    * Provides access to app's constants from the legacy module registry.
    */
   val constants: ConstantsInterface?
@@ -158,12 +171,18 @@ class AppContext(
   /**
    * Provides access to the file system service
    */
-  val filePermission by lazy { servicesProvider.filePermission() }
+  val filePermission: FilePermissionService
+    get() = service<FilePermissionService>() ?: throw IllegalStateException(
+      "FilePermissionService is not registered in the ServicesRegistry."
+    )
 
   /**
    * Provides access to the scoped directories from the legacy module registry.
    */
-  private val appDirectories by lazy { servicesProvider.appDirectories() }
+  private val appDirectories: AppDirectoriesService
+    get() = service<AppDirectoriesService>() ?: throw IllegalStateException(
+      "AppDirectoriesService is not registered in the ServicesRegistry."
+    )
 
   /**
    * A directory for storing user documents and other permanent files.
@@ -187,24 +206,6 @@ class AppContext(
    * Provides access to the image loader from the legacy module registry
    */
   val imageLoader: ImageLoaderInterface?
-    get() = legacyModule()
-
-  /**
-   * Provides access to the camera view manager from the legacy module registry
-   */
-  val camera: CameraViewInterface?
-    get() = legacyModule()
-
-  /**
-   * Provides access to the font manager from the legacy module registry
-   */
-  val font: FontManagerInterface?
-    get() = legacyModule()
-
-  /**
-   * Provides access to the task manager from the legacy module registry
-   */
-  val taskManager: TaskManagerInterface?
     get() = legacyModule()
 
   /**
@@ -311,7 +312,12 @@ class AppContext(
     hostWasDestroyed = true
   }
 
-  internal fun onActivityResult(activity: Activity, requestCode: Int, resultCode: Int, data: Intent?) {
+  internal fun onActivityResult(
+    activity: Activity,
+    requestCode: Int,
+    resultCode: Int,
+    data: Intent?
+  ) {
     activityResultsManager.onActivityResult(requestCode, resultCode, data)
     registry.post(
       EventName.ON_ACTIVITY_RESULT,
@@ -335,7 +341,9 @@ class AppContext(
   @UiThread
   fun <T : View> findView(viewTag: Int): T? {
     val reactContext = hostingRuntimeContext.reactContext ?: return null
-    return UIManagerHelper.getUIManagerForReactTag(reactContext, viewTag)?.resolveView(viewTag) as? T
+    return UIManagerHelper
+      .getUIManagerForReactTag(reactContext, viewTag)
+      ?.resolveView(viewTag) as? T
   }
 
   internal fun dispatchOnMainUsingUIManager(block: () -> Unit) {