+++ packages/expo-router/src/toolbar/elements.tsx
@@ -15,6 +15,12 @@ export interface ToolbarMenuProps {
   accessibilityHint?: string;
   children?: React.ReactNode;
   /**
+   * An optional subtitle for the menu. Does not appear on `inline` menus.
+   *
+   * @see [Apple documentation](https://developer.apple.com/documentation/uikit/uimenuelement/subtitle) for more information.
+   */
+  subtitle?: string;
+  /**
    * If `true`, the menu item will be displayed as destructive.
    *
    * @see [Apple documentation](https://developer.apple.com/documentation/uikit/uimenu/options-swift.struct/destructive) for more information.
@@ -47,7 +53,8 @@ export interface ToolbarMenuProps {
   inline?: boolean;
   /**
    * If `true`, the menu will be displayed as a palette.
-   * This means that the menu will be displayed as one row
+   * This means that the menu will be displayed as one row.
+   * The `elementSize` property is ignored when palette is used, all items will be `elementSize="small"`. Use `elementSize="medium"` instead of `palette` to display actions with titles horizontally.
    *
    * > **Note**: Palette menus are only supported in submenus.
    *
@@ -88,6 +95,15 @@ export interface ToolbarMenuProps {
    * @default 'plain'
    */
   variant?: 'plain' | 'done' | 'prominent';
+  /**
+   * The preferred size of the menu elements.
+   * `elementSize` property is ignored when `palette` is used.
+   *
+   * @see [Apple documentation](https://developer.apple.com/documentation/uikit/uimenu/preferredelementsize) for more information.
+   *
+   * @platform iOS 16.0+
+   */
+  elementSize?: 'auto' | 'small' | 'medium' | 'large';
 }
 
 /**
@@ -113,6 +129,7 @@ export const ToolbarMenu: React.FC<ToolbarMenuProps> = ({
   palette,
   inline,
   hidden,
+  subtitle,
   title,
   destructive,
   children,
@@ -120,6 +137,7 @@ export const ToolbarMenu: React.FC<ToolbarMenuProps> = ({
   tintColor,
   variant,
   style,
+  elementSize,
 }) => {
   const identifier = useId();
   const validChildren = Children.toArray(children).filter(
@@ -142,10 +160,12 @@ export const ToolbarMenu: React.FC<ToolbarMenuProps> = ({
       hidden={hidden}
       icon={sf}
       destructive={destructive}
+      subtitle={subtitle}
       accessibilityLabel={accessibilityLabel}
       accessibilityHint={accessibilityHint}
       displayAsPalette={palette}
       displayInline={inline}
+      preferredElementSize={elementSize}
       tintColor={tintColor}
       titleStyle={titleStyle}
       barButtonItemStyle={variant === 'done' ? 'prominent' : variant}
@@ -440,8 +460,8 @@ export interface ToolbarViewProps {
  *       placeholderTextColor={Color.ios.placeholderText}
  *     />
  *   </Toolbar.View>
- *   <Toolbar.View separateBackground style={{ width: 32, height: 32 }}>
- *     <Pressable onPress={handlePress}>
+ *   <Toolbar.View separateBackground>
+ *     <Pressable style={{ width: 32, height: 32 }} onPress={handlePress}>
  *       <SymbolView name="plus" size={22} />
  *     </Pressable>
  *   </Toolbar.View>